py3: update CloseableQueue and tests to enable use under Python 3
In `CloseableQueue.py`, the imports are now wrapped in a try/except
block.  The new Python 3 imports are done in the `try` clause and the
original Python 2 imports are left intact under the `except` clause.

To update the tests, used a [recent version of test_queue.py][1] as a
reference to find out how the mainline copy evolved since the creation
of CloseableQueue.  I integrated the changes that allowed the test
cases to run without crashing.

[1]: https://github.com/python/cpython/blob/v3.7.2/Lib/test/test_queue.py